refactoring-client-file-locations
Refactoring of the Client files locations has been done
