+++ src/Components/Airport/AirportModal.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Modal from 'react-modal';
 import { Button, TextField, Box, Typography, FormHelperText } from '@mui/material';
 
-import { apiService } from '../../apiService';
+import { apiService } from '../../util/apiService';
 
 Modal.setAppElement('#root');
