+++ src/Components/Employee/EmployeeList.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Button, Typography, Box, Container, Modal, Paper } from '@mui/material';
 import DataTable from '../../util/DataTable';
 import EmployeeModal from './EmployeeModal';
-import { apiService } from '../../apiService';
+import { apiService } from '../../util/apiService';
 
 const EmployeeList = () => {
     const [employees, setEmployees] = useState([]);
