+++ src/Components/Employee/EmployeeModal.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import Modal from 'react-modal';
 import { Button, TextField, Box, Typography } from '@mui/material';
-import { apiService } from '../../apiService';
+import { apiService } from '../../util/apiService';
 
 Modal.setAppElement('#root');
