--- src/Components/Users/Login.module.css
@@ -1,101 +1,0 @@
-.login-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  padding: 1rem;
-}
-
-.login-card {
-  background: white;
-  padding: 2rem;
-  border-radius: 12px;
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.2);
-  text-align: center;
-  width: 100%;
-  max-width: 500px;
-  height: 550px;
-  opacity: 0.9;
-}
-
-.login-header h2 {
-  margin: 0 0 0.5rem 0;
-  font-size: 1.5rem;
-  color: #333;
-}
-
-.login-header p {
-  margin: 0 0 50px 0px;
-  color: #666;
-}
-
-.form-group {
-  margin-bottom: 1.5rem;
-  text-align: left;
-}
-
-.form-group label {
-  display: block;
-  margin-bottom: 0.5rem;
-  color: #555;
-}
-
-.input-container {
-  display: flex;
-  align-items: center;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  padding: 0.5rem;
-}
-
-.input-icon {
-  margin-right: 0.5rem;
-  color: #888;
-}
-
-.form-group input {
-  width: 100%;
-  border: none;
-  outline: none;
-  padding: 0.75rem;
-}
-
-.login-button {
-  background: #2575fc;
-  color: white;
-  padding: 0.75rem;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  width: 100%;
-  font-size: 1rem;
-  transition: background 0.3s;
-}
-
-.login-button:hover {
-  background: #1a5cbf;
-}
-.success-message {
-  color: green;
-  margin-top: 20px;
-}
-
-.error-message {
-  color: red;
-  margin-top: 20px;
-}
-
-@media (max-width: 480px) {
-  .login-card {
-    padding: 1.5rem;
-  }
-
-  .form-group-- input {
-    padding: 0.5rem;
-  }
-
-  .login-button {
-    padding: 0.5rem;
-    font-size: 0.875rem;
-  }
-}
